Delete TipCalc.java
not needed/not related to project.
From Instructor:
"Why/how is TipCalc related to everything else? Will it assign a random tip? It looks like meal_amount and tip_percentage are unique to that class. Therefore, I assume it will calculate the tip (and isn't really a descendent of the Number class)."